Add files via upload
This will Allow you to get rid of pesky objects getting in the way of your edits.
I'm surprised to not find this already in blender.

basically it checks every half second which mode you're in and hides every other object not excluded from the auto hide. it stores the status of the objects and puts them back after.